# executor: vectorize the merge join executor by `vecGroupChecker`