@qw4990 2020-01-19
Hi @catror , I add a new test for the vectorized merge join in this PR https://github.com/pingcap/tidb/pull/14519, please merge it into your PR to check this PR's correctness at your convenience.
And also please fix this CI problem in the `check_dev_2` test:
```
[2020-01-17T09:51:51.002Z] FAIL: physical_plan_test.go:784: testPlanSuite.TestAggToCopHint
[2020-01-17T09:51:51.002Z] 
[2020-01-17T09:51:51.002Z] physical_plan_test.go:824:
[2020-01-17T09:51:51.002Z]     c.Assert(planString, Equals, output[i].Best, comment)
[2020-01-17T09:51:51.002Z] ... obtained string = "TableReader(Table(t))"
[2020-01-17T09:51:51.002Z] ... expected string = "TableReader(Table(t)->HashAgg)->HashAgg->HashAgg"
[2020-01-17T09:51:51.002Z] ... case:2 sql:select /*+ AGG_TO_COP(), HASH_AGG(), USE_INDEX(t) */ distinct a from t group by a
```