@catror 2020-02-03
> @catror Friendly ping, please resolve conflicts.

Done.